editor/libeditor/tests/test_bug483651.html

<!DOCTYPE HTML>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=483651
-->

<head>
  <title>Test for Bug 483651</title>
  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>

<body onload="doTest();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=483651">Mozilla Bug 483651</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>

  <pre id="test">
    <script type="application/javascript">

      /** Test for Bug 483651 **/

      SimpleTest.waitForExplicitFinish();

      function doTest() {
        var t1 = $("t1");
        var editor = null;

        if (t1 instanceof Components.interfaces.nsIDOMNSEditableElement)
          editor = t1.editor;

        ok(editor, "able to get editor for the element");
        t1.focus();
        synthesizeKey("A", {});
        synthesizeKey("VK_BACK_SPACE", {});

        try {
          // Was the trailing br removed?
          is(editor.documentIsEmpty, true, "trailing <br> correctly removed");
        } catch (e) {
          ok(false, "test failed with error "+e);
        }
        SimpleTest.finish();
      }
   </script>
  </pre>

  <textarea id="t1" rows="2" columns="80"></textarea>
</body>
</html>
